feature(train): refactor feature engineering file into smaller ones
